Conflicts:
	java/src/probcog/rosie/RosieGUI.java

+++ java/src/probcog/navigation/MonteCarloPlanner.java
@@ -334,10 +334,10 @@ public class MonteCarloPlanner
 
         // Initialize planning components
         HashMap<String, TypedValue> params = new HashMap<String, TypedValue>();
-        params.put("side", new TypedValue((byte)-1));
+        params.put("side", new TypedValue(-1));
         params.put("distance", new TypedValue((double)0.75));
         controls.add(new FollowWall(params));
-        params.put("side", new TypedValue((byte)1));
+        params.put("side", new TypedValue(1));
         controls.add(new FollowWall(params));
 
         // Initialize wavefront planner used for rough distance
